+++ ark_auth/src/client/async_impl/key.rs
@@ -6,13 +6,13 @@ use futures::Future;
 impl AsyncClient {
     pub fn key_create(
         &self,
-        is_active: bool,
+        is_enabled: bool,
         name: &str,
         service_id: Option<&str>,
         user_id: Option<&str>,
     ) -> impl Future<Item = CreateResponse, Error = Error> {
         let body = CreateBody {
-            is_active,
+            is_enabled,
             name: name.to_owned(),
             service_id: service_id.map(|x| x.to_owned()),
             user_id: user_id.map(|x| x.to_owned()),