+++ ark_auth/src/client/async_impl/user.rs
@@ -6,13 +6,13 @@ use futures::Future;
 impl AsyncClient {
     pub fn user_create(
         &self,
-        is_active: bool,
+        is_enabled: bool,
         name: &str,
         email: &str,
         password: Option<&str>,
     ) -> impl Future<Item = CreateResponse, Error = Error> {
         let body = CreateBody {
-            is_active,
+            is_enabled,
             name: name.to_owned(),
             email: email.to_owned(),
             password: password.map(String::from),