+++ ark_auth/src/client/sync_impl/service.rs
@@ -26,12 +26,12 @@ impl SyncClient {
 
     pub fn service_create(
         &self,
-        is_active: bool,
+        is_enabled: bool,
         name: &str,
         url: &str,
     ) -> Result<CreateResponse, Error> {
         let body = CreateBody {
-            is_active,
+            is_enabled,
             name: name.to_owned(),
             url: url.to_owned(),
         };