+++ ark_auth/src/client/sync_impl/user.rs
@@ -5,13 +5,13 @@ use crate::server::route::user::{CreateBody, CreateResponse};
 impl SyncClient {
     pub fn user_create(
         &self,
-        is_active: bool,
+        is_enabled: bool,
         name: &str,
         email: &str,
         password: Option<&str>,
     ) -> Result<CreateResponse, Error> {
         let body = CreateBody {
-            is_active,
+            is_enabled,
             name: name.to_owned(),
             email: email.to_owned(),
             password: password.map(String::from),